Updated version of Texture functions

+++ SimdSse2Texture_amd64.go
@@ -9,10 +9,10 @@ import (
 
 
 //go:noescape
-func _SimdSse2TextureBoostedSaturatedGradient(src unsafe.Pointer, srcStride, width, height uint64, saturation, boost uint8, dx unsafe.Pointer, dxStride uint64, dy unsafe.Pointer, dyStride uint64)
+func _SimdSse2TextureBoostedSaturatedGradient(src unsafe.Pointer, srcStride, width, height uint64, saturation, boost uint64/*uint8*/, dx unsafe.Pointer, dxStride uint64, dy unsafe.Pointer, dyStride uint64)
 
 //go:noescape
-func _SimdSse2TextureBoostedUv(src unsafe.Pointer, srcStride, width, height uint64, boost uint8, dst unsafe.Pointer, dstStride uint64)
+func _SimdSse2TextureBoostedUv(src unsafe.Pointer, srcStride, width, height uint64, boost uint64/*uint8*/, dst unsafe.Pointer, dstStride uint64)
 
 //go:noescape
 func _SimdSse2TextureGetDifferenceSum(src unsafe.Pointer, srcStride, width, height uint64, lo unsafe.Pointer, loStride uint64, hi unsafe.Pointer, hiStride uint64, sum unsafe.Pointer)
@@ -21,13 +21,13 @@ func _SimdSse2TextureGetDifferenceSum(src unsafe.Pointer, srcStride, width, heig
 func _SimdSse2TexturePerformCompensation(src unsafe.Pointer, srcStride, width, height uint64, shift int, dst unsafe.Pointer, dstStride uint64)
 
 //
-func SimdSse2TextureBoostedSaturatedGradient(src View, saturation, boost uint8, dx, dy View) {
+func SimdSse2TextureBoostedSaturatedGradient(src View, saturation, boost uint64/*uint8*/, dx, dy View) {
 
 	_SimdSse2TextureBoostedSaturatedGradient(src.GetData(), uint64(src.GetStride()), uint64(src.GetWidth()), uint64(src.GetHeight()), saturation, boost, dx.GetData(), uint64(dx.GetStride()), dy.GetData(), uint64(dy.GetStride()))
 }
 
 //
-func SimdSse2TextureBoostedUv(src View, boost uint8, dst View) {
+func SimdSse2TextureBoostedUv(src View, boost uint64/*uint8*/, dst View) {
 
 	_SimdSse2TextureBoostedUv(src.GetData(), uint64(src.GetStride()), uint64(src.GetWidth()), uint64(src.GetHeight()), boost, dst.GetData(), uint64(dst.GetStride()))
 }
